+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning v2.0.0](http://semver.org/).
 
+## [1.1.1] - 2026-05-09
+### Added
+- Support for class metadata factories chaining
+
+### Fixed
+- Bug with missing class metadata factory configuration in the configurator
+
 ## [1.1.0] - 2026-05-09
 ### Added
 - Support for inherited (through JOINED) Doctrine ORM entities
@@ -27,3 +34,4 @@ This project adheres to [Semantic Versioning v2.0.0](http://semver.org/).
 [1.0.1]: https://github.com/cryonighter/formula-doctrine/tree/v1.0.1
 [1.0.2]: https://github.com/cryonighter/formula-doctrine/tree/v1.0.2
 [1.1.0]: https://github.com/cryonighter/formula-doctrine/tree/v1.1.0
+[1.1.0]: https://github.com/cryonighter/formula-doctrine/tree/v1.1.1